# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /auth/register/creator:
    post:
      operationId: AuthController_registerCreator
      summary: Регистрация креатора
      description: "Первый вход: создаёт Account (email/password) и первый профиль
        User с ролью CREATOR. Выставляет cookies access-token и refresh-token.
        Email должен быть уникальным."
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterCreatorDto"
      responses:
        "201":
          description: Account и профиль созданы, пользователь авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionUserResponse"
        "409":
          description: Email уже зарегистрирован
      tags: &a3
        - auth
  /auth/register/company:
    post:
      operationId: AuthController_registerCompany
      summary: Регистрация компании
      description: "Первый вход: создаёт Account (email/password) и первый профиль
        User с ролью COMPANY. Выставляет cookies. Email должен быть уникальным."
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterCompanyDto"
      responses:
        "201":
          description: Account и профиль созданы, пользователь авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionUserResponse"
        "409":
          description: Email уже зарегистрирован
      tags: *a3
  /auth/login:
    post:
      operationId: AuthController_login
      summary: Вход
      description: Аутентификация по email и password Account. Активирует первый
        доступный профиль (или тот, что был при последнем switch). Выставляет
        cookies. rememberMe=true — refresh на 30 дней, иначе 7 дней.
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
      responses:
        "200":
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionUserResponse"
        "401":
          description: Неверный email или пароль
      tags: *a3
  /auth/profiles:
    get:
      operationId: AuthController_listProfiles
      summary: Список доступных профилей
      description: "Все профили User, к которым текущий Account имеет доступ: свои
        (OWNER) и чужие через invite. Используйте для переключателя профилей на
        фронте. Поле isActive — текущий профиль в JWT."
      parameters: []
      responses:
        "200":
          description: Массив профилей с displayName, role, membershipRole
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProfileSummaryResponse"
      tags: *a3
      security:
        - access-token: []
  /auth/switch-profile:
    post:
      operationId: AuthController_switchProfile
      summary: Переключить активный профиль
      description: Меняет активный User в JWT (sub). Нужен userId из GET
        /auth/profiles. После переключения все запросы (чат, PATCH профиля) идут
        от имени выбранного профиля. Обновляет cookies.
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchProfileDto"
      responses:
        "200":
          description: Профиль переключён, новые cookies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionUserResponse"
        "403":
          description: Нет membership к этому профилю
      tags: *a3
      security:
        - access-token: []
  /auth/invites:
    post:
      operationId: AuthController_createInvite
      summary: Пригласить менеджера к существующему профилю
      description: "Отправляет письмо на email приглашённого. userId —
        **существующий** профиль, которым вы управляете. Требуется роль OWNER
        или ADMIN на этом профиле. Роли invite: ADMIN, EDITOR, VIEWER (не
        OWNER). Приглашённый должен зарегистрироваться/войти своим email и
        принять invite."
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviteDto"
      responses:
        "201":
          description: Приглашение создано, письмо отправлено
        "403":
          description: Недостаточно прав (нужен OWNER/ADMIN на профиле)
      tags: *a3
      security:
        - access-token: []
  /auth/invites/accept:
    post:
      operationId: AuthController_acceptInvite
      summary: Принять приглашение
      description: Принимает invite по token из письма (ссылка
        /invites/accept?token=...). Email залогиненного Account должен совпадать
        с email в приглашении. После accept профиль появится в GET
        /auth/profiles.
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptInviteDto"
      responses:
        "200":
          description: Приглашение принято, membership создан
        "400":
          description: Недействительный или просроченный token
        "403":
          description: Email аккаунта не совпадает с email в приглашении
      tags: *a3
      security:
        - access-token: []
  /auth/memberships/{id}:
    delete:
      operationId: AuthController_revokeMembership
      summary: Отозвать доступ к профилю
      description: Удаляет membership другого человека к профилю, которым вы
        управляете. id — membershipId из GET /auth/profiles. Нельзя отозвать
        OWNER. Требуется OWNER или ADMIN на профиле.
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: Доступ отозван
        "403":
          description: Недостаточно прав
      tags: *a3
      security:
        - access-token: []
  /auth/recovery-password:
    post:
      operationId: AuthController_recoveryPassword
      summary: Запрос сброса пароля
      description: Отправляет письмо со ссылкой сброса, если Account с таким email
        существует. Всегда возвращает успех (защита от перебора email).
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestPasswordResetDto"
      responses:
        "200":
          description: Сообщение об отправке письма (независимо от наличия email)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
      tags: *a3
  /auth/reset-password:
    post:
      operationId: AuthController_resetPassword
      summary: Установить новый пароль
      description: Меняет password Account по token из письма. Token передаёт фронт из
        query-параметра URL письма.
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordDto"
      responses:
        "200":
          description: Пароль изменён
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "400":
          description: Недействительный или просроченный token
      tags: *a3
  /auth/verify-password:
    post:
      operationId: AuthController_verifyPassword
      summary: Проверить текущий пароль
      description: Проверяет пароль залогиненного Account. Используйте перед сменой
        пароля в настройках. Требует авторизацию (cookie access-token).
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPasswordDto"
      responses:
        "200":
          description: Пароль верный
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyPasswordResponseDto"
        "401":
          description: Не авторизован
        "403":
          description: Неверный пароль
      tags: *a3
      security:
        - access-token: []
  /auth/refresh:
    post:
      operationId: AuthController_refresh
      summary: Обновить access token
      description: Читает refresh-token из cookie, выдаёт новую пару токенов. Вызывать
        при истечении access-token или proактивно с фронта.
      parameters: []
      responses:
        "200":
          description: Токены обновлены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionUserResponse"
        "401":
          description: Неверный или устаревший refresh token (нужен перелогин)
      tags: *a3
...
info:
  title: Project Back API
  description: >-
    # Project Back API


    Backend для marketplace creator ↔ company. Документация для фронтенда и
    AI-ассистентов.


    ## Модель данных


    | Сущность | Что это |

    |----------|---------|

    | **Account** | Login: email + password. Один человек = один Account. |

    | **User** | Публичный профиль: CREATOR или COMPANY. При регистрации
    создаётся один User на Account. |

    | **Membership** | Доступ Account к User: OWNER / ADMIN / EDITOR / VIEWER. |

    | **Invite** | Приглашение другому человеку управлять **существующим**
    профилем. |


    **Активный профиль** — тот, чей `userId` в JWT (`sub`). Переключение: `POST
    /auth/switch-profile`.


    ## Cookies и CORS


    - После login/register/refresh выставляются httpOnly cookies:
    `access-token`, `refresh-token`.

    - Все запросы с `credentials: 'include'`.

    - CORS origin: `CORS_ORIGIN` (обычно http://localhost:3000).


    ## Сценарии


    ### Первый вход

    1. `POST /auth/register/creator` или `/auth/register/company` — создаёт
    Account + первый профиль.

    2. `POST /auth/login` — вход по email/password.


    ### Совместное управление (менеджер)

    1. Владелец: `POST /auth/invites` — email менеджера + `userId`
    **существующего** профиля + роль.

    2. Менеджер: регистрация/логин **своим** email → `POST
    /auth/invites/accept`.

    3. Менеджер: `GET /auth/profiles` → `POST /auth/switch-profile`.


...

...

...


...

...

...


...

...

...

...

...
servers:
...
components:
  securitySchemes:
    cookie:
      type: apiKey
      in: cookie
      name: access-token
  schemas:
    ContactItemDto:
      type: object
      properties:
        type:
          type: string
          enum:
            - phone
            - telegram
            - whatsapp
            - instagram
            - youtube
            - vk
            - website
            - other
          example: phone
          description: "Тип контакта: phone, telegram, instagram и т.д."
        value:
          type: string
          example: "+79991234567"
          description: Номер телефона или URL соцсети
        label:
          type: string
          example: WhatsApp
          description: "Подпись: «Рабочий», «Личный Telegram»"
      required:
        - type
        - value
...
    RegisterCreatorDto:
      type: object
      properties:
        contacts:
          description: "Контакты: телефоны и ссылки на соцсети"
          type: array
          items:
            $ref: "#/components/schemas/ContactItemDto"
        phone:
          type: string
          example: "+79991234567"
          description: Основной телефон
        location:
          type: string
          example: Moscow
        avatar:
          type: string
          example: https://example.com/avatar.png
        bio:
          type: string
          example: Short bio
        aboutMe:
          type: string
          example: More about me
        email:
          type: string
          example: creator@example.com
        password:
          type: string
          example: password123
          minLength: 8
        name:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
      required:
        - email
        - password
        - name
        - lastName
    AuthSessionUserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum:
            - CREATOR
            - COMPANY
        membershipRole:
          type: string
          enum:
            - OWNER
            - ADMIN
            - EDITOR
            - VIEWER
        isVerified:
          type: boolean
          description: Профиль проверен платформой (бейдж верификации)
        isEmailConfirmed:
          type: boolean
          description: Email профиля (User.email) подтверждён
      required:
        - id
        - role
        - membershipRole
        - isVerified
        - isEmailConfirmed
    RegisterCompanyDto:
      type: object
      properties:
        contacts:
          description: "Контакты: телефоны и ссылки на соцсети"
          type: array
          items:
            $ref: "#/components/schemas/ContactItemDto"
        phone:
          type: string
          example: "+79991234567"
          description: Основной телефон
        location:
          type: string
          example: Moscow
        avatar:
          type: string
          example: https://example.com/avatar.png
        bio:
          type: string
          example: Short bio
        aboutMe:
          type: string
          example: More about me
        email:
          type: string
          example: company@example.com
        password:
          type: string
          example: password123
          minLength: 8
        companyName:
          type: string
          example: Acme Inc.
      required:
        - email
        - password
        - companyName
    LoginDto:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
        password:
          type: string
          example: password123
        rememberMe:
          type: boolean
          default: false
          description: "Запомнить меня: refresh-токен и cookie живут 30 дней вместо 7"
      required:
        - email
        - password
    ProfileSummaryResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum:
            - CREATOR
            - COMPANY
        displayName:
          type: string
          example: John Doe
        avatar:
          type: string
          nullable: true
        membershipRole:
          type: string
          enum:
            - OWNER
            - ADMIN
            - EDITOR
            - VIEWER
        membershipId:
          type: string
          format: uuid
        isActive:
          type: boolean
        isVerified:
          type: boolean
          description: Профиль проверен платформой (бейдж верификации)
        isEmailConfirmed:
          type: boolean
          description: Email профиля (User.email) подтверждён
      required:
        - userId
        - role
        - displayName
        - avatar
        - membershipRole
        - membershipId
        - isActive
        - isVerified
        - isEmailConfirmed
    SwitchProfileDto:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          description: userId профиля из GET /auth/profiles, на который переключиться
      required:
        - userId
    CreateInviteDto:
      type: object
      properties:
        email:
          type: string
          example: manager@example.com
          description: Email приглашённого (должен совпасть с Account при accept)
        userId:
          type: string
          format: uuid
          description: userId существующего профиля, к которому даётся доступ
        role:
          type: string
          enum:
            - ADMIN
            - EDITOR
            - VIEWER
          description: ADMIN — полный доступ + invite; EDITOR — редактирование; VIEWER —
            только чтение
      required:
        - email
        - userId
        - role
    AcceptInviteDto:
      type: object
      properties:
        token:
          type: string
          description: Token из ссылки в письме-приглашении
      required:
        - token
    RequestPasswordResetDto:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
      required:
        - email
    ResetPasswordDto:
      type: object
      properties:
        token:
          type: string
          description: Token из ссылки в письме сброса пароля
        newPassword:
          type: string
          example: newPassword123
          minLength: 8
          description: Новый пароль Account (мин. 8 символов)
      required:
        - token
        - newPassword
    VerifyPasswordDto:
      type: object
      properties:
        password:
          type: string
          example: currentPassword123
          description: Текущий пароль Account для проверки перед сменой
      required:
        - password
    VerifyPasswordResponseDto:
      type: object
      properties:
        valid:
          type: boolean
          example: true
      required:
        - valid
...